JournAI/requirements.txt
accelerate==1.5.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bitsandbytes @ git+https://github.com/TimDettmers/bitsandbytes.git@b86ff64bf57aa800dd6ab1419d6038d71915e3d3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.2.1
datasets==3.4.1
deepspeed==0.16.4
dill==0.3.8
diskcache==5.6.3
einops==0.8.1
fastapi==0.115.12
filelock==3.18.0
frozenlist==1.5.0
fsspec==2024.12.0
h11==0.16.0
hjson==3.1.0
huggingface-hub==0.29.3
idna==3.10
Jinja2==3.1.6
llama_cpp_python==0.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.2.0
multiprocess==0.70.16
networkx==3.4.2
ninja==1.11.1.4
numpy==2.2.4
packaging==24.2
pandas==2.2.3
peft==0.15.0
propcache==0.3.1
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==19.0.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.3
scipy==1.15.2
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
sympy==1.13.1
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.50.2
trl==0.16.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.35.0
xxhash==3.5.0
yarl==1.18.3
